Docker permission denied problem
I have used quorum maker on couple of machines for development purposes and it is installed without any problems. However, I have faced the same problem in each of them in regards to Docker giving permission denied problem. This usually happens if the current user has access limitation to docker engine. The simple solution is mentioned on the following link:
https://techoverflow.net/2017/03/01/solving-docker-permission-denied-while-trying-to-connect-to-the-docker-daemon-socket/
We just need to give access to the user with the following command:
sudo usermod -a -G docker $USER

Please review the suggestion and provide your feedback.